feat(aim): unify trim algebra - ray+laser adopt the model quat compose
